Refactor existing functions and add resetDisplay()
Refactored appendExpression() to check for solvable expressions if
there is a trailing operator in expression (ex. 4++ or 5+4-). This is
so only one expression will be solved at a time.

Refactored evaluateExpression() to ignore the first operator in
expression is there are multiple, so negative numbers won’t break the
calculator. Also to add event listener to numbers buttons so
clicking one after expression is evaluated resets display and current
expression stored.

Refactored populateDisplay() so it detects if there is an unfinished
expression.

Added resetDisplay() as a callback function, it resets display if a
number button is clicked after expression is evaluated with equals key
ONLY if there isn’t an active operator.